fix impossibility to invoke equake with no graphical frames.
When there are no graphical frames on a monitor we can't determine
explicitly what's the monitor name. That means we can't tell Emacs to
create a frame on this exact monitor. Before this commit, we tried to
create frames in a terminal of a selected frame which led to errors
when the terminal was text-only.

No we just create some graphical frame (with an utility of '(getenv
"DISPLAY")') and determine a monitor name after the fact