apply several heuristics to guess a DISPLAY value
Since there is no way (known to me) to reliably get a display
value when it's not set in the environment, we have to use a
heuristic.  Every time a graphical frame is opened we store its
display value in `equake-persistent-display-file'.  As long, as
we don't change a display, this value remains valid and we can
safely and properly launch Equake at all times.

This approach does not eliminate the problem completely, since
there is still a corner case when a user tries to invoke Equake
on a new display before any other graphical frames have been
launched.

So, in case it fails we go through widely-used DISPLAY
values (e.g. :0, :1, w32) and try to connect to those. A user can
adjust this list of values.