fix recording of Equake buffer history
- addresses https://gitlab.com/emacsomancer/equake/-/issues/26

"When =equake-use-frame-hide= is =nil=, then history of Equake frame does not seem to work properly. Instead of actual Equake buffers being restored, another buffer appears in the Equake frame. I think this is because frame destruction in some contexts leaves behind records of dead buffers which get improperly indexed. A potential fix seems to be to check whether a buffer is actually an Equake buffer before storing a record for it. The drawback is that if a user had a non-Equake buffer open in Equake when they 'hide' the Equake frame, it will not re-appear automatically when Equake is re-invoked. Still, defaulting to restoring only Equake buffers seems like a better outcome than an arbitrary Emacs buffer appearing when Equake is re-opened."